+++ httpizza/src/main/java/sexy/code/Callback.java
@@ -1,8 +1,8 @@
 package sexy.code;
 
-public interface Callback<T> {
+public interface Callback {
 
-    void onResponse(Response<T> response);
+    void onResponse(Response response);
 
     void onFailure(Throwable t);
 }