+++ httpizza/src/main/java/sexy/code/Response.java
@@ -1,16 +1,15 @@
 package sexy.code;
 
-
 import java.util.List;
 import java.util.Map;
 
-public class Response<T> {
+public class Response {
 
     private final HttpResponse origin;
 
-    private final T body;
+    private final ResponseBody body;
 
-    Response(HttpResponse origin, T body) {
+    Response(HttpResponse origin, ResponseBody body) {
         this.origin = origin;
         this.body = body;
     }
@@ -27,15 +26,19 @@ public class Response<T> {
         return origin.isSuccess();
     }
 
-    public int getCode() {
+    public int statusCode() {
         return origin.getCode();
     }
 
-    public Map<String, List<String>> getHeaders() {
+    public String message() {
+        return origin.getMessage();
+    }
+
+    public Map<String, List<String>> headers() {
         return origin.getHeaders();
     }
 
-    public T getBody() {
+    public ResponseBody body() {
         return body;
     }
 }