+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+## Version 0.2.0
+
+_2016-02-09_
+
+* Removed request and response body conversion for now
+* Added OkHttp MediaType and tests
+* Changed API a bit
+
 ## Version 0.1.0
 
 _2016-02-01_